feat: 个人中心头像上传完善

+++ ticho-rainbow-app/src/main/java/top/ticho/rainbow/interfaces/facade/FileInfoController.java
@@ -98,7 +98,7 @@ public class FileInfoController {
         @ApiImplicitParam(value = "是否限制", name = "limit"),
     })
     @GetMapping("getUrl")
-    public Result<String> getUrl(Long id, Integer expire, Boolean limit) {
+    public Result<String> getUrl(Long id, Long expire, Boolean limit) {
         return Result.ok(fileInfoService.getUrl(id, expire, limit));
     }
